draw index file
